backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-auth==2.31.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.122.0
python-dotenv==1.0.1
google-generativeai==0.4.0
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9
